+++ CHANGELOG.md
@@ -5,6 +5,11 @@ All notable changes to Calendar Map Filter (CMF) will be documented in this file
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.4.9] - 2025-11-3
+
+- Fixed timezone bug in foopee.ts eventsource.
+- use waitUntil() for umamiServer() to not block response 
+
 ## [0.4.8] - 2025-10-25
 
 - added `noExpire=1` option to `/api/events` to prevent cache from expiring. Use with `skipCache=1`